requirements.txt
alembic==1.16.4
appdirs==1.4.4
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cssselect==1.3.0
exceptiongroup==1.3.0
fake-useragent==2.2.0
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
psycopg2-binary==2.9.10
pyee==11.1.1
PyJWT==2.10.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
requests-html==0.10.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.43
tomli==2.2.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==1.26.20
w3lib==2.3.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.23.0
